Form signature (#17525)
* Scaffold form signature component and test

* Reduce scope

Specifically, don't include any content. It'll be up to engineers to
plug it in themselves and use the signature component in there as
needed.

* Render custom label

* React element label

* Remove "on behalf of"

After reducing the scope like this, it's easy to just tack it on as
needed in the `CustomComponent`.

* Update example

* Use TextInput

* Re-add default props

* Categorize tests and default required: false

* Add checkbox and validation

* Required validation

* Run custom validations

* Simplify the test

* Change required error message

Because the validation doesn't actually check that the name matches
anything, it only checks to make sure the box is filled out. It may not
be an especially useful validation, but we can't generically check that
the name matches what's in the form data because we don't know what the
data structure looks like for any given form.

What we _can_ do, however, is provide a validator function creator that
accepts the first and last names and returns the validator function.

* Call onSectionComplete

* Update tests with required error message content

* Add invalid test

* Add prop descriptions

* setFormData default prop

* setFormData unit test

* Refactor maybeCompleteSection() to useEffect()

* Remove validators placeholder test describe

* Update default error message text

* Add aria-labelledby prop

* Remove the commented props

* Add Attestation components

This wraps the FormSignature and passes an ariaLabelledBy prop.

* Update the description of the FormSignature

* Add Attestation description

* a11y fixes

* fix typo and add comment

Co-authored-by: Christopher Valarida <chris.valarida@adhocteam.us>
Co-authored-by: Katryn Bowman <katy.bowman@adhocteam.us>