Add ESLint rule for table component migration (#20633)
* Add rough draft of table migration rule

* Bump component-library to get helper function

* Handle sorting props

This doesn't interpret the props in any way. If the `currentSort` prop
is preset, we add `sort-column="0" descending` so that the engineer has
an idea of how the props work and they can make their own adjustments
from there. After all, this functionality is just a suggestion.

* Update comment

* Update component-library to latest version

This is to get the sorting functionality for table