- How to apply buttons hidden behind feature flag. (#22085)
Root cause:

The affected buttons were conditioned to the MEB feature flag being false. When we turned on MEB feature flag so How To Apply allowed a claimant to reach MEB when they answered the right set of questions, the side effect was that the affected buttons were now hidden.

The logic for gating these buttons based on the MEB feature flag is incorrect.

Fix summary:

These buttons should not care about the MEB feature flag. Remove the condition that hides the affected buttons based on the feature flag.  I've tested locally and the buttons show up as expected. The MEB button still shows up as expected with the change.