Mhv 33532 landing page sort (#22088)
* [#MHV-33532] message sorting feature on InboxListView component

sorting messages on a component level, given component accepts an array of message data as a prop. Additionally, modifying mockdata sent dates temporarily until api calls are established

* removing unused unit test

Header component is no longer used. Somehow in between the merges, this unit test got left behind on this branch, but is already killed on the main. But this unit test fails a PR check since the component is missing