Use apiRequest in COE form.
Context:

During testing, we noticed that our auth tokens were expiring midway through completing the COE form. This is because we were using fetchAndUpdateSessionExpiration in our fetchInProgressForm action and in DocumentUploader/submit.js, instead of apiRequest. See Slack thread: https://dsva.slack.com/archives/CBU0KDSB1/p1665003605654539?thread_ts=1665002706.486319&cid=CBU0KDSB1. Unlike the other request methods, apiRequest renews expired tokens. This PR is a big step towards using apiRequest throughout the COE form.

This is one of two issues blocking us from moving into QA.

Solution:

Updating DocumentUploader/submit.js was a simple task. Updating fetchInProgressForm was more complicated. This is because 1) fetchInProgressForm's error handling relies heavily on the http status of the response, but 2) apiRequest did not expose the http status on error. To be specific, when apiRequest receives an error not caused by an expired auth token, it would reject the promise with the json response body, but not the response object and the status code contained therein.

This PR adds a rejectWithFullResponse argument to apiRequest. When true, request errors not caused by an expired auth token cause the promise to be rejected with the Response object. This allows fetchInProgressForm to consume the http status of the response, just as it did before.

With respect to successful requests, because apiRequest resolves promises with the response body, rather than the Response object itself, I have had to re-write fetchInProgressForm accordingly. While reviewing this PR, you'll notice some strange-ness around Promise rejection. Namely, that Promises are not rejected with Errors, and that Promises can be rejected with more than one kind of type. While this is not ideal, 1) fetchInProgressForm and apiRequest were already doing this, 2) my PR is focused on an urgent bugfix, not a larger re-write of fetchInProgressForm and apiRequest, and 3) apiRequest was built on the assumption that most http error statuses will, unlike fetch, reject Promises rather than resolve them (meaning, I am forced to surface http error statuses through reject).

I could re-write apiRequest to handle errors differently, but because it is so widely used, it's difficult for me to justify the kind of re-write that would put me on what could be a lengthy test-fixing expedition.

Outstanding issues:

There are other COE endpoints (GETs) not addressed by this PR, like GET /v0/coe/document_download/{id} and GET /v0/coe/download_coe. We consume these endpoints through hrefs and not fetch requests, so it's not immediately clear to me how to avoid the expired auth token issue for these endpoints. I can explore this later.