41046 update military service branches in 526ez (#22476)
This PR requests military service branches from lighthouse's API, caches them in sessionStorage, and uses these cached branches to populate dropdowns throughout 526ez. If for whatever reason the request to lighthouse fails, we use a hardcoded list of branches returned by lighthouse at the time of this PR. This PR also includes migrations/transformations that pluralize previously singular branch names (e.g. Reserve -> Reserves) when forms are prefilled, populated with in progress form data, or submitted to the backend. The frontend and backend are able to digest these updated branches because of updates to vets-json-schema in department-of-veterans-affairs/vets-json-schema#725 and department-of-veterans-affairs/vets-json-schema#726.

Co-authored-by: Robin Garrison <robin.garrison@adhocteam.us>