Use local components & styles instead of site-wide assets in pre-check-in app (#22667)
* Allow to opt-out of the web-components chunk

This is only for local scaffolding, and will depend on a change in the
content-build repo to add the appropriate JSON value to the app's
registry entry.

* Add the pre-release with the new build

* Manually define the telephone web component

This also imports the component library CSS file with the CSS variables
since we no longer have the wc-loader file on the page. More web
component definitions will follow

* Switch import strategy to a single function

I think importing a function (eventually several) from a single path and
then calling it is a bit friendlier than having to import a Component
class from its own unique path, and _then_ define it with the browser's
customElements registry

* Define remaining web components used in this app

* Reorder the component definitions alphabetically

* Update name of registry value

In response to the content-build change:

https://github.com/department-of-veterans-affairs/content-build/pull/1358/commits/b1125e69861e0d62ac41909d79a53523d22e4ec4

* Begin work on creating the minimal stylesheet

* Switch to stable release

* Regroup some sass dependencies

* Add some more Sass modules after trial and error

* Remove commented out core sass import

* Conditionally apply minimal stylesheet to non-prod

Eventually we'll want to revert this once the opt-out for the site-wide
stylesheet actually impacts prod. I'm doing this for now so that we
don't have the site-wide stylesheet _and_ the minimal stylesheet as part
of the app bundle, resulting in a larger overall load.