Add unit tests
Added comment explaining why we were leaving in commented code. Also
added a test for checking that the loading element appears if the
feature flag is loading.

Note that the coverage is a little wonky - the linter requires (rightly)
an ReactWrapper.unmount() for every enzyme mount() call, but for some
reason, calling `unmount()` causes coverage to suffer.