Fixed false positives
Previously was using enzyme to check for presence of components. This
led to a bunch of false positives. If I removed the render of a
subcomponent in the parent component, the test would still pass. I
figured this was due to some async stuff in the `useEffect` hook, so I
read up a little bit and realized that the react-testing-library works
well with async things. I switched to use react-testing-library and am
now confident that things are being tested correctly.