49594 ddl unit tests (#23052)
* Adding first unit test

* Adding unit tests

* Add unit tests for YourClaimLetters

* Reorganize components of YourClaimLetters

* Remove lint disable

* Add unit tests

Added comment explaining why we were leaving in commented code. Also
added a test for checking that the loading element appears if the
feature flag is loading.

Note that the coverage is a little wonky - the linter requires (rightly)
an ReactWrapper.unmount() for every enzyme mount() call, but for some
reason, calling `unmount()` causes coverage to suffer.

* Fixed false positives

Previously was using enzyme to check for presence of components. This
led to a bunch of false positives. If I removed the render of a
subcomponent in the parent component, the test would still pass. I
figured this was due to some async stuff in the `useEffect` hook, so I
read up a little bit and realized that the react-testing-library works
well with async things. I switched to use react-testing-library and am
now confident that things are being tested correctly.

Co-authored-by: jerekshoe <jerek.shoemaker@adhocteam.us>