Fixing res.json() being called incorrectly
A bug was discovered where fetchInProgressForm was calling res.json() inside of a console.log and then attempted to read it again as part of a return. Because res.json() only reads and outputs the json version of the response body 'once', the returned value was empty and calling errors.