Don't allow non-US countries in COE form.
During pilot testing we discovered that LGY doesn't accept non-US, non-military-base addresses. So, we will not allow such addresses in the COE form. We expect this will be fine, because it seems unlikely that many vets will be applying for a COE, for a home in the US, while residing primarily outside of the US. Actually, it may not be allowed. See https://mortgagesolutions.net/2022/10/19/use-va-mortgage-loan-outside-u-s/.

The form system, however, expects a country field within the address form, so we are adding an additional parameter `onlyAllowUSAddresses` that sets the country to the US and disables the input.