Remove `willRefinance` parameter from payload to LGY. (#23254)
LGY's API docs indicate that they don't accept this parameter, we never sent the parameter to begin with, and although we used to collect it on the front end (i.e. "Do you want to refinance this loan?"), now that we are sending a loan intent for each loan (See https://github.com/department-of-veterans-affairs/vets-api/pull/11666 and https://github.com/department-of-veterans-affairs/vets-website/pull/23249), this question is unnecessary because the loan intent answers the question.