[53062] Require LOA and EDIPI for COE Form (#23308)
After launching COE to 25% of vets, we discovered that the COE form and COE status page don't validate whether a user is LOA3 or has an EDIPI -- both of which are necessary to submit a COE form. This PR enforces both requirements before user can interact with the COE form or COE status page. See companion back-end PR: https://github.com/department-of-veterans-affairs/vets-api/pull/11732.