Add Google Analytics logging to message display
Adds a recordEvent call to log the appearance of the error message and dynamically includes which data was missing. May consider a better way to include the list of missing identifiers in the analytics log so it is easier to query on in the future.

Fix typo on participantId key in test and prop types