Fixing NOD useEffect infinite loop issue.
We encountered an issue where `useEffect` would be called over and over because an `issuesNeedUpdating` check was guaranteed to return true if a Veteran had contestable issues with a decision date older than one year.

For context, `issuesNeedUpdating` was intended to check whether contestable issue data loaded from the API was different from contestable issue data that already existed in `formData`. If the data was different, we would run a series of data normalization and filtering functions that would, among other things, ensure that Veterans could not see contestable issues with a decision date older than a year. These checks are somewhat expensive, so we didn't want to run them unless we needed to. Which is why the `issuesNeedUpdating` check was there.

However, the `issuesNeedUpdating` function was comparing unfiltered data from the API (i.e. all contested issues, regardless of decision date) with filtered data from `formData` (only containing issues with a decision date within the past year). This meant that when a Veteran had issues with a decision date older than one year, `issuesNeedUpdating` was guaranteed to return true, no matter what, because the filtered data would necessarily be different from the unfiltered data.

The fix, introduced here, is to compare _filtered_ data from the API with _filtered_ data from `formData`.