Improve performance and add code comments.
As to performance, we've split out all the code relating to contested issues into its own `useEffect`. This is because the code is a bit expensive, and not something we want to run every single time `formData` changes. Also, in `setFormData` we are setting `contestedIssues` to the result of `getEligibleContestableIssues` (the filtering function), which itself calls `processContestableIssues` (the data normalization function). Before, we were calling `processContesableIssues` on the result of `getEligibleContestableIssues`, which was redundant.