Remove disabled flaky test for claims form keyboard nav
As per https://depo-platform-documentation.scrollhelp.site/developer-docs/currently-disallowed-cypress-test-specs-on-va-gov, this has been automatically silenced from running during CI because it fails intermittently and needs to be rewritten. As of 11/06/23, merging will be blocked if this test isn't rewritten or removed.

We're opting to remove it for now; a ticket has been created to rewrite this test when we have time to dedicate to it:  https://app.zenhub.com/workspaces/disability-benefits-experience-team-carbs-6470c8bfffee9809b2634a52/issues/gh/department-of-veterans-affairs/va.gov-team/68612