Remove feature flag toggles
Removes the front end feature flags I initially implemented here alongside the back end changes in https://github.com/department-of-veterans-affairs/vets-api/pull/14858. I decided it would simplify the code and release rollout to simply check if the new missing identifiers metadata was present (meaning the back end feature toggle was on), and make the component rendering dependent on that instead of checking a flag that is unecessary in this case.