526: Implement more accurate "We're missing some information" messages (#25428)
* Implement 526 missing identifiers warning component

Creates a basic component and unit tests to tell the user they are missing certain identifiers we require before starting the Form 526 flow, and dynamically renders which identifiers are missing so they can share that information with a contact center

This is intended to replace all of the the inconsistent and inaccurate message components in src/applications/disability-benefits/all-claims/containers/MissingServices.jsx

* Refactor Missing526Identifiers containerm tests

Cleans up the formatting and rendering code for Missing526Identifiers to make it more logically organized. Addresses problems with tests that seemed to be working before; ended up having to use a regexp match on the container content for my message paragraph; using typical React Testing Library functions such as getByText is problematic and causes confusing failures because it doesn't support web components, which va-alert is (it's part of the VA component library)

* Add Google Analytics logging to message display

Adds a recordEvent call to log the appearance of the error message and dynamically includes which data was missing. May consider a better way to include the list of missing identifiers in the analytics log so it is easier to query on in the future.

Fix typo on participantId key in test and prop types

* Refactor Form526 to render new error messages

Replaces the existng Missing Services component renders in Form526EZApp.jsx with the new Missing526Identifiers component. This was done to standardize the messsaging during these situations and to provide the user with exactly which data is missing to share with the contact center.

Adds default claims arg to the test page in Form526EZApp.unit.spec.jsx. These styles of tests are now deprecated in our system, but I kept my new tests in the same style for consistency.

* Refactor list formatting for older browser compatibility

Adds a new utility function to our claims utilities for formatting a list as a human readable string with the correct commas and use of 'and'.

Replaces my initial implementation of this in the Missing526Identifiers endpoint which used the Intl.ListFormat function:
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl/ListFormat

As this is a relatively new feature of JavaScript not supported by older browsers (I was told Safari 9 for example, which many veterans are still on, frequently raises errors when we use these features), it's better to use the more vanilla, classic JavaScript approach

* Feature gate new and old messaging behavior

Adds back the existing messages behavior I had removed in an earlier commit and wraps the new and old behavior in a feature flag.

* Add empty string safeguard to title formatter

The default argument to this titleLowerCase function is an empty string, however the destructuring below it will leave the firstChar undefined and would error if we tried to lower case it

* Refactor Alert to use React children pattern

* Improve language in Missing526Identifiers spec

* Remove feature flag toggles

Removes the front end feature flags I initially implemented here alongside the back end changes in https://github.com/department-of-veterans-affairs/vets-api/pull/14858. I decided it would simplify the code and release rollout to simply check if the new missing identifiers metadata was present (meaning the back end feature toggle was on), and make the component rendering dependent on that instead of checking a flag that is unecessary in this case.

* Revert accidental removal of important docstring from existing code

* Add slot attr to h2 in VA alert

* Use existing utility for formatting lists

I wasn't aware of this readable list utility when I made my own, so deleting my implementation and relying on this existing helper.

* Fix docstring typo

* PR Feedback: Add v3 uswds styling tag to alert message

Adds the uswds tag to the alert poriton of the new More Info Needed error message for styling consistency