21a into arp - basic correction to expected reducer structure (#30927)
Fixing this error unlocks the next error which I believe exposes a more meaningful incompatibility that SIP experiences when the the app's state has not been set up by the platform user component. The key idea here is that in progress forms are fundamentally connected to a user. to integrate in progress forms into our app we'll need to introduce and expose user state that in progress forms understands