Update feature toggle to be stored in formData rather than sessionStorage
Storing the feature toggle in sessionStorage does not work in production because by the time the toggle value is stored and then retrieved the form has already loaded. This approach stores the feature toggle in the formData and uses depends to check that formData and conditionally show the correct content.