Fix storing of feature toggle in formData and onFormLoad
Aligns with the implementation found in src/applications/appeals/995/containers/App.jsx. Now the feature toggle setup successfully allows rolling out and rolling back the feature toggle. The only remaining issue is that if they are currently in the form when the feature toggle is toggled they will get a error message and need to refresh their browser.