Update addDisabilitiesPrevious page path to be less  confusing to users
This is in alignment with the advice from Information Architecture. First they advised to just alter the page the user is on: "Change the URL slug for the exact page, not a parent page - so if you are changing the page that lives at "add" in this URL /benefit/apply-form123/dependents/add, then change the /add slug, not the /dependents/ slug.". Second they advised to "Avoid any words that could create confusion or alarm with the user that they may not be on the right page.  In other words, I would stay away from using terms like "old" or "archive", and instead use something that it less noticeable like just appending a number to the URL." The content of this page is the 2nd version of the content hence the /add-2.