Switch new content path to /add-3 and old content path to /add
This will allow us to rollout the feature more slowly as 95% of users will go to /add and only 5% to /add-3 for a extended time to allow seeing forms get fully submitted, and in the meantime the Google Analytics will be only slightly disrupted. The add-3 references that this is the 3rd version of the page. When it is assured that /add-3 content is working well a change will be merged to move the content from /add-3 to /add.