97067 disability benefits nb workflow choice page skeleton (#33655)
* WORKING COMMIT

* WORKING COMMIT cleaning up changes

* WORKING COMMIT Still trying to figure out tests

* Get view field working with current web component patterns

* Clean up latest iteration of tests to see why not working

* Remove conditional routing utility unit test

Unit testing this functionality was proving problematic with Sinon; it was decided to wait to test the conditional routing based on worlflow choice in proper end to end integration tests instead. These will be implemented later in the project

* Remove placeholder mental health info page it was replaced with implementation already

* Fix view property requirement syntax to get tests passing again

* Fix syntax on property key in utility function for workflow choice

* Add heading tag with form name

* Add page description

* Add choices description

* Add selection labels

* Fix formatting on header

* Add collapsable additional info on questions to be asked

* Fix broken routing

Checking the form data in the utility functions to see what workflow choice was selected got altered by mistake during the master merge

Additionally, the initial implementation of the routing for the consent page was unaware of the isCompletingForm0781 check (opt in), so it needed to be added for that page

* Fix broken test

* Implement dynamic mental health condition list

Adds a previously missed requirement to dynamically list the mental health conditions selected on the mental health screener page when making selection on what 0781 flow the user wants to complete

* Refactor conditions list for new spec

Updates the conditions list display on the 0781 workflow choice page to list all new disabiites claimed by the user per the latest spec

* Add space before conditions list

Fixes the formatting in the conditions list display on the 0781 workflow choice page to include a space befor ethe list

* Use curly format apostrophe

* Update existing workflow choice copy for new revisions

* Replace existing additional info dropdown with mental health condition accordion card

* Fix outdated content import

* Change dynamic condition list to ul element

Veteran-selected conditions that are listed on the Form 0781 workflow choice page can sometimes including commas, this can be confusing in a paragraph element when there are multiple conditions, so we are switching to a bulleted list

* PR feedback: fix docstring typo

* Remove defensive guard clause on conditions list

I had initially implemented this page in a way where some of the content would still render if for some reason the list of newly claimed conditions was missing. The user should not be able to get to this page if that has happened; therfore it is preferable for this page to blow up instead of rendering incomplete content

* Add more robust test for error message

Add test for error message content