Conditions 671 Update Feature Toggle for Platform ArrayField and Autocomplete Progressive Rollout (#34038)
* Update feature toggle to be stored in formData rather than sessionStorage

Storing the feature toggle in sessionStorage does not work in production because by the time the toggle value is stored and then retrieved the form has already loaded. This approach stores the feature toggle in the formData and uses depends to check that formData and conditionally show the correct content.

* Update Autocomplete addDisabilities tests for Platform ArrayField

Platform ArrayField 'Edit' button is a va-button whereas All Claims ArrayField 'Edit' button is not. va-button requires a test util since it is a web component and text cannot be retrieved with getByText.

* Add check fo feature toggle being undefined to fix failing tests

* Fix duplicate property names test by adding to the ignore array

* Update e2e testing to update var name & test with toggle on

* Fix storing of feature toggle in formData and onFormLoad

Aligns with the implementation found in src/applications/appeals/995/containers/App.jsx. Now the feature toggle setup successfully allows rolling out and rolling back the feature toggle. The only remaining issue is that if they are currently in the form when the feature toggle is toggled they will get a error message and need to refresh their browser.

* Resolve failing App tests by adding formData to check and to dep array

* Fix tests from merging main with ArrayField VaButton update

This PR updated the 'update' and 'remove' buttons to VaButtons: https://github.com/department-of-veterans-affairs/vets-website/pull/34024/files.

* Fix e2e tests from merging main with ArrayField VaButton update

This PR updated the 'update' and 'remove' buttons to VaButtons: https://github.com/department-of-veterans-affairs/vets-website/pull/34024/files\.

* Fix null checks for formData in mapStateToProps to prevent potential runtime errors

From PR feedback: https://github.com/department-of-veterans-affairs/vets-website/pull/34038#discussion_r1916958897

* Update addDisabilitiesPrevious page path to be less  confusing to users

This is in alignment with the advice from Information Architecture. First they advised to just alter the page the user is on: "Change the URL slug for the exact page, not a parent page - so if you are changing the page that lives at "add" in this URL /benefit/apply-form123/dependents/add, then change the /add slug, not the /dependents/ slug.". Second they advised to "Avoid any words that could create confusion or alarm with the user that they may not be on the right page.  In other words, I would stay away from using terms like "old" or "archive", and instead use something that it less noticeable like just appending a number to the URL." The content of this page is the 2nd version of the content hence the /add-2.

* Switch new content path to /add-3 and old content path to /add

This will allow us to rollout the feature more slowly as 95% of users will go to /add and only 5% to /add-3 for a extended time to allow seeing forms get fully submitted, and in the meantime the Google Analytics will be only slightly disrupted. The add-3 references that this is the 3rd version of the page. When it is assured that /add-3 content is working well a change will be merged to move the content from /add-3 to /add.

* Update cypress helper in alignment with url path update

* Add issue url to TODO and update onFormLoaded TODO from PR feedback