Hide legacy Form 0781 flow when new flow is enabled
The hasNewPtsdDisability utility function gates the first page of the existing Form 0781 flow for Form 526. When the syncModern0781Flow flipper is set to true, we will display a new version of the PTSD section later in the Form 526 flow. When this new version is visible, hasNewPtsdDisability should always return false so the legacy flow is hidden.

Adds additional tests to three functions that call hasNewPtsdDisability: isAnswering781Questions, isAnswering781aQuestions and needsToEnter781a. These all serve as gates for exposing different legacy pages, so testing them all is an extra layer of saftey to ensure none of the legacy PTSD flow is visible when syncModern0781Flow is true