Refactor trigger for entering new Form 0781 flow
Removes the formData?.mentalHealth?.conditions check from showForm0781Pages, the primary gate function for displaying the new version of the 0781 flow.

This conditions array is populated by user selections on the mental health screener page, which is getting removed. So the showForm0781Pages check needs to ignore that data.

Note, because there is a default value of true for isClaimingNew if !hasRatedDisabilities is true, we need to add the !isClaimingIncrease clause to be absolutely sure the user is making a new claim and not claiming an increase, because that is the only situation where we would display the Form 0781 flow