101360 disability benefits nb hide legacy ptsd flow when new flow enabled (#34515)
* Hide legacy Form 0781 flow when new flow is enabled

The hasNewPtsdDisability utility function gates the first page of the existing Form 0781 flow for Form 526. When the syncModern0781Flow flipper is set to true, we will display a new version of the PTSD section later in the Form 526 flow. When this new version is visible, hasNewPtsdDisability should always return false so the legacy flow is hidden.

Adds additional tests to three functions that call hasNewPtsdDisability: isAnswering781Questions, isAnswering781aQuestions and needsToEnter781a. These all serve as gates for exposing different legacy pages, so testing them all is an extra layer of saftey to ensure none of the legacy PTSD flow is visible when syncModern0781Flow is true

* Ensure new 0781 flow flipper is false in 0781 e2e test

Adds cy intercept to ensure 0781 legacy and 0781a pages work correctly when the new 0781 flow flipper toggle is turned off

* Fix feature toggle name in mocked backend response in E2E test

* Reordering steps in setup to see if it fixes the tests

---------

Co-authored-by: jerekshoe <jerek.shoemaker@adhocteam.us>