Simplify new condition option and fix update bug
When rated disability is selected the option was "Change to new condition", but when clicked the option immediately updates to "Edit new condition". This is because the logic for the option checks isRatedDisability which changes. This solution simplifies the space and solves this bug by always making the option "New condition".