Resolves sentry/platform-web-m4 (#35474)
- issue URL: sentry.vfs.va.gov/organizations/vsp/issues/?query=PLATFORM-WEB-M4

vets_sip_list_item_missing_info error is reported to Sentry by isSIPEnabledForm
fn via loading the <ApplicationsInProgress /> component at the /my-va path. The
list of saved forms for the current user is compared against the "presentable"
forms the application knows about. The form is presentable if it has a title
and a URL. vets_sip_list_item_missing_info error is reported to Sentry if
either the form title or URL is not known for a particular saved form.

Based on the frequency of occurance (2.6m events for 328k users), this signal
doesn't seem useful. This signal simply indicates that a user has a form in
their history that isn't supported by the <ApplicationsInProgress /> component.