Align loader between downtime and FF loader
Align message across DowntimeNotification, feature flag, and RequiredLoginView loaders to be Loading your information... RequiredLoginView cannot be altered without changing platform code so the message aligns with its default message for now. Align and add set-focus to DowntimeNotification and feature flag loader to inform the user that loading is occurring.