Create add_team_reviewer_cop_checker.yml (#36705)
This work will add a new workflow to vets-website Pull Requests. 

When a PR has changes requested by the CoP engineer, the CoP is removed as a reviewer from the Pull Request. This will workflow is designed to fix the issue.

The workflow will trigger on a submitted pull request with changes requested. It will take the reviewers GitHub username and reference the CoP GitHub team. If the reviewer is a member of the va-platform-cop-frontend, the CoP will be added back to the PR as a reviewer. 

More detailed information on this can be found within this GitHub ticket: https://github.com/department-of-veterans-affairs/va.gov-team/issues/107277