Remove duplicate file input component behind feature flag (#37094)
Previously for the cstFriendlyEvidenceRequests feature toggle the VaFileInput label text was planned to differ between feature toggled enabled and disabled. As a result  the VaFileInput was duplicated to be in both the enabled and disabled slots. However, currently the VaFileInput should be the same between enabled and disabled so the duplicate is not necessary. Note that the VaFileInputs were accidentally different in that the enabled (which was never in prod) had the uswds prop. It should not have the uswds prop as that would make the VaFileInput based on uswds v3 which would break the implementation.