feat: Update Docker for Review Instances (#39446)
* Add insecure flag to curl for AWS CLI download

The '-k' flag was added to the curl command to allow connections to servers with insecure SSL certificates when downloading the AWS CLI installer. This may be necessary in environments where certificate verification fails.

* Disable npm strict-ssl in Dockerfile

Added a command to set npm's strict-ssl to false, which may help resolve SSL issues when installing packages in certain environments.

* Update docker-compose.review.yml