534 ez Chapter 4 - Update Previous Marriages (Spouse & Veteran) flow (#40248)
* update previous question name to spouseMarriages

* Add conditional page flow for spouse and veteran marriages

Updated veteranChildren and veteranMarriagesPages configs to conditionally display pages based on claimantRelationship and hadPreviousMarriages. This ensures that marriage-related pages are only shown to spouses who indicate previous marriages.

* Update array paths for marriage pages

Changed arrayPath and related references from 'previousMarriages' to 'spouseMarriages' in previousMarriagesPages.js and to 'veteranMarriages' in veteranMarriagesPages.js. This improves clarity and consistency in form data structure for spouse and veteran previous marriages.

* include unit tests for previous marriages update